### 2023-07-08 #5629 @editorialbot
[PRE REVIEW]: LBR-Stack: ROS 2 and Python Integration of KUKA FRI for Med and IIWA Robots
**Submitting author:** @mhubii (<a href="http://orcid.org/0000-0003-4603-6773">Martin Huber</a>)
**Repository:** https://github.com/lbr-stack/lbr_stack_doc
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
