### 2023-08-04 #5629 @adi3
@destogl @SteveMacenski @traversaro @SeungBack - Are any of you available to review this ROS paper?
