### 2023-08-04 #5629 @SteveMacenski
While I don't argue against the utility of what they've put together, I don't see any research impact / value that would be meet the journal's requirements. It appears to me to be another wrapper on a company's SDK to expose to ROS. I'd be happy to review the paper once its generated, but I'm not sure this work is appropriate for any journal publication.
