### 2023-08-05 #5629 @mhubii
Thank you for the early feedback @SteveMacenski , we are trying our best to find a solution to the paper generation.

In the meantime, the paper is accessible in the documenting repository https://github.com/lbr-stack/lbr_fri_ros2_stack_doc/actions/runs/5749570650.

The compilation does not work through @editorialbot  right now because the documentation is separate from the software repository. We do want to apologize for the inconvenience.
