### 2023-08-06 #5629 @mhubii
Hi @adi3 , is it possible to re-submit? Not sure how to get the @editorialbot to change the repository.

The submission is regarding the [lbr-stack](https://github.com/lbr-stack/), which also includes ROS bindings but much more. Think that is what caused the confusion for @SteveMacenski. This is our mistake. Really sorry.
