+++ doc/source/conf.py
@@ -75,10 +75,10 @@ pygments_style = "sphinx"
 
 html_theme = "pydata_sphinx_theme"
 html_theme_options = {
-    "github_url": "https://github.com/kCL-BMEIS/lbr_fri_ros2_stack/tree/humble",
+    "github_url": "https://github.com/lbr-stack/lbr_fri_ros2_stack/tree/humble",
     "switcher": {
         "version_match": "humble",
-        "json_url": "https://raw.githubusercontent.com/KCL-BMEIS/lbr_fri_ros2_stack_doc/humble/doc/source/switcher.json",
+        "json_url": "https://raw.githubusercontent.com/lbr-stack/lbr_fri_ros2_stack_doc/humble/doc/source/switcher.json",
     },
     "navbar_start": ["navbar-logo", "version-switcher"],
 }
