+++ doc/source/index.rst
@@ -21,4 +21,4 @@
    :hidden:
    :caption: Getting Involved
 
-   Source Code <https://github.com/kCL-BMEIS/lbr_fri_ros2_stack/tree/humble>
+   Source Code <https://github.com/lbr-stack/lbr_fri_ros2_stack/tree/humble>
