### 2023-10-10 #5629 @CameronDevine
I was another author of the CoreRobotics paper/software and a reviewer for one of the papers listed above. I am willing to review this paper if you are still searching for reviewers.
