### 2023-11-22 df011c2
Add citation info #20
