### 2023-12-27 #6138 @CameronDevine
@adi3 Before I get too far into this review, I would like to make sure we are all in agreement about the scope of this review. In the past, when I have both submitted and reviewed JOSS papers, the codebase and paper have been in a single repository. In this case, the codebase appears to be spread among multiple repositories all housed in a common GitHub organization. Therefore, should @bmagyar and I review all the repositories in the GitHub organization?

One exception to this rule would, I assume, be the [lbr-stack/pymoveit2](https://github.com/lbr-stack/pymoveit2) repository which is a fork of another repository.

Finally, the checklist asks if the source code is available at https://github.com/lbr-stack/lbr_stack_doc, which it is not. Does this need to be changed?
