### 2024-01-19 #6138 @arfon
> @arfon some valid points by @CameronDevine above. How should we go about handling multi-repo codebases?

This does happen sometimes, and if you believe the structure of the project is reasonable, then it's acceptable for the review to be across multiple codebases.
