### 2024-01-30 lbr-stack/lbr_stack_doc#21 @CameronDevine
[JOSS Review] Add Missing DOI
As part of the JOSS Review, openjournals/joss-reviews#6138, please add this missing DOI number for _Towards MRI-based autonomous robotic US acquisitions: a first feasibility study_: `10.1109/tmi.2016.2620723`.
