### 2024-02-03 62ebe0b
fixes [JOSS Review] Add Missing DOI #21
