+++ paper/paper.bib
@@ -48,14 +48,14 @@
 }
 
 @article{iiwa_stack,
-  title={Towards MRI-based autonomous robotic US acquisitions: a first feasibility study},
   author={Hennersperger, Christoph and Fuerst, Bernhard and Virga, Salvatore and Zettinig, Oliver and Frisch, Benjamin and Neff, Thomas and Navab, Nassir},
-  journal={IEEE transactions on medical imaging},
+  journal={IEEE Transactions on Medical Imaging}, 
+  title={Towards MRI-Based Autonomous Robotic US Acquisitions: A First Feasibility Study}, 
+  year={2017},
   volume={36},
   number={2},
-  pages={538--548},
-  year={2017},
-  publisher={IEEE}
+  pages={538-548},
+  doi={10.1109/TMI.2016.2620723}
 }
 
 @inproceedings{fri,
