### 2024-02-05 #6138 @mhubii
thank you for the feedback @CameronDevine. Hope things are going well!

As rightfully indicated, the license is now OSI compliant. Copyright notices where third-party software was used are included.

Looking forward for further feedback so we can work through this review :)
