### 2024-03-08 #6138 @mhubii
what about `vincentberenz`, who was initially keen to review?

https://github.com/openjournals/joss-reviews/issues/6138#issuecomment-1849486352
