+++ paper/paper.bib
@@ -104,5 +104,6 @@
   title={Reducing the barrier to entry of complex robotic software: a moveit! case study},
   author={Coleman, David and Sucan, Ioan and Chitta, Sachin and Correll, Nikolaus},
   journal={arXiv preprint arXiv:1404.3785},
-  year={2014}
+  year={2014},
+  doi={10.6092/JOSER_2014_05_01_p3}
 }
