+++ doc/source/conf.py
@@ -78,7 +78,7 @@ html_theme_options = {
     "github_url": "https://github.com/lbr-stack/lbr_fri_ros2_stack/tree/humble",
     "switcher": {
         "version_match": "humble",
-        "json_url": "https://raw.githubusercontent.com/lbr-stack/lbr_fri_ros2_stack_doc/humble/doc/source/switcher.json",
+        "json_url": "https://raw.githubusercontent.com/lbr-stack/lbr_stack_doc/main/doc/source/switcher.json",
     },
     "navbar_start": ["navbar-logo", "version-switcher"],
 }
