+++ doc/source/conf.py
@@ -77,10 +77,6 @@ html_theme = "pydata_sphinx_theme"
 html_logo = "../img/logo.png"
 html_theme_options = {
     "github_url": "https://github.com/lbr-stack/lbr_fri_ros2_stack/tree/humble",
-    "switcher": {
-        "version_match": "humble",
-        "json_url": "https://raw.githubusercontent.com/lbr-stack/lbr_stack_doc/main/doc/source/switcher.json",
-    },
     "navbar_start": ["navbar-logo", "version-switcher"],
 }
 html_context = {
