+++ doc/source/conf.py
@@ -76,6 +76,7 @@ pygments_style = "sphinx"
 
 html_theme = "pydata_sphinx_theme"
 html_logo = "../img/logo.png"
+html_extra_path = ["lbr_fri_ros2_stack"] # for images in README.md
 html_theme_options = {
     "github_url": "https://github.com/lbr-stack/lbr_fri_ros2_stack/tree/humble",
     "navbar_start": ["navbar-logo", "version-switcher"],
