+++ doc/source/index.rst
@@ -3,7 +3,8 @@
    You can adapt this file completely to your liking, but it should at least
    contain the root `toctree` directive.
 
-.. mdinclude:: lbr_fri_ros2_stack/README.md
+.. include:: lbr_fri_ros2_stack/README.md
+   :parser: myst_parser.sphinx_
 
 .. toctree::
    :hidden:
