+++ doc/source/index.rst
@@ -13,14 +13,14 @@
    lbr_fri_ros2_stack/lbr_fri_ros2_stack/doc/lbr_fri_ros2_stack
    lbr_fri_ros2_stack/lbr_bringup/doc/lbr_bringup
    lbr_fri_ros2_stack/lbr_demos/doc/lbr_demos
-   lbr_fri_ros2_stack/lbr_fri_ros2/doc/lbr_fri_ros2
-   fri/doc/fri
    lbr_fri_ros2_stack/docker/doc/docker
    lbr_fri_ros2_stack/lbr_moveit_config/doc/lbr_moveit_config
    docs/additional_resources
-
-.. toctree::
-   :hidden:
-   :caption: Getting Involved
-
+   fri/doc/fri
+   lbr_fri_ros2_stack/lbr_fri_ros2/doc/lbr_fri_ros2
    Source Code <https://github.com/lbr-stack/lbr_fri_ros2_stack/tree/humble>
+   Changelog <https://github.com/lbr-stack/lbr_fri_ros2_stack/releases>
+
+
+
+
