### 2024-04-13 lbr-stack/lbr_stack_doc#23 @CameronDevine
Statement of Need Missing
As part of the JOSS review (openjournals/joss-reviews#6138), a Statement of Need should be included in the documentation as well as the paper.
