### 2024-04-13 lbr-stack/lbr_stack_doc#24 @CameronDevine
Community Guidelines Missing
As part of the JOSS review (openjournals/joss-reviews#6138) there must be clearly stated community guidelines, and I haven't been able to find any. I think these guidelines could either be added to your readthedocs.io documentation, or included in each repository. For your reference, the following is required:

> Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support
