+++ .readthedocs.yml
@@ -13,6 +13,7 @@ build:
   jobs:
     post_install:
       - vcs import doc/source --input https://raw.githubusercontent.com/lbr-stack/lbr_fri_ros2_stack/humble/lbr_fri_ros2_stack/repos.yaml
+      - unzip doc/source/fri/FRI-Client-SDK_Cpp.zip -d doc/source/fri/FRI-Client-SDK_Cpp
 
 python:
    install:
