+++ doc/source/lbr_fri_idl.rst
@@ -1,0 +1,2 @@
+.. include:: lbr_fri_idl/README.md
+   :parser: myst_parser.sphinx_
