+++ doc/source/fri.rst
@@ -1,0 +1,2 @@
+.. include:: fri/README.md
+    :parser: myst_parser.sphinx_
