### 2024-05-15 #6138 @dmronga
@adi3  I have a question regarding the functionality of the software, which claims to work on 4 different KUKA LBR robots as one of the main contributions. Obviously, this cannot be tested without having the actual hardware available (although I am intending to do a test on at least the LBR iiwa 14, which we have in our lab). I assume checking if the Gazebo simulation is working should be enough in this case?
