### 2024-05-15 #6138 @dmronga
@mhubii I found some minor [issues](https://github.com/lbr-stack/lbr_stack_doc/issues) (apart from those already mentioned by previous reviewer). Feel free to have a look any time
