### 2024-05-27 #6138 @CameronDevine
> @CameronDevine can I ask you to please clear out the last few items in your checklist? Would appreciate it!

I will try to complete as much as I can today. I am also still waiting on a few issues to be closed before my review is complete. Specifically:

* lbr-stack/pyFRI#15
* lbr-stack/pyFRI#14
* lbr-stack/lbr_stack_doc#24
* lbr-stack/lbr_stack_doc#23
