### 2024-06-11 be8cecb
added venv instructions (#25)
