### 2024-06-11 7e2db7f
added link to paper (#23)
