### 2024-06-11 f61b58d
removed fri_vendor
