### 2024-06-11 60c637c
Create CODE_OF_CONDUCT.md, https://github.com/lbr-stack/lbr_stack_doc/issues/24
