### 2024-06-12 lbr-stack/lbr_stack_doc#24 @CameronDevine
Just as a reminder, the community guidelines should be included in each repository, or added to the documentation. Let me know if you would like me to open issues in the other repositories.
