### 2024-06-12 lbr-stack/lbr_stack_doc#24 @mhubii
contributing:

| Repository | CONTRIBUTING.md | CODE_OF_CONDUCT.md |
| ---------------- | -----------------------------| -- |
|  `lbr_stack_doc`  |  [ ] [link]()  | [x] [link](https://github.com/lbr-stack/lbr_stack_doc/blob/main/CODE_OF_CONDUCT.md)|
|  `fri`  | [ ] [link]()   | [ ] [link]() |
|  `pyFRI`  | [x] [link](https://github.com/lbr-stack/pyFRI/blob/main/.github/CONTRIBUTING.md)    | [x] [link](https://github.com/lbr-stack/pyFRI/blob/main/CODE_OF_CONDUCT.md)|
|  `lbr_fri_ros2_stack`  | [x] [link](https://github.com/lbr-stack/lbr_fri_ros2_stack/blob/humble/.github/CONTRIBUTING.md)    | [x] [link](https://github.com/lbr-stack/lbr_fri_ros2_stack/blob/humble/CODE_OF_CONDUCT.md) |
|  `lbr_fri_idl`  | [ ] [link]()    | [ ] [link]() |
