### 2024-06-12 101cae8
Create CONTRIBUTING.md, https://github.com/lbr-stack/lbr_stack_doc/issues/24
