+++ .github/CONTRIBUTING.md
@@ -1,0 +1,16 @@
+The [LBR-Stack](https://github.com/lbr-stack/) thrives to build a friendly community for `KUKA` users to foster knowledge exchange! :)
+
+We **encourage** everyone to open issues for questions / problems / features. For additional information, we would like to provide some links below.
+
+### Knowledge Sources
+- [Open issues](https://github.com/lbr-stack/lbr_stack_doc/issues)
+- [Closed issues](https://github.com/lbr-stack/lbr_stack_doc/issues?q=is%3Aissue+is%3Aclosed)
+- [Read the Docs](https://lbr-stack.readthedocs.io/en/latest/)
+
+## Contributing
+### LBR Stack Doc
+Contributions are vital to this project. If you want to contribute a feature / fix proceed as follows
+
+1. Open an issue: Explain the issue and your solution
+2. Fork this repository
+3. Create a [pull request](https://github.com/lbr-stack/lbr_stack_doc/pulls)
