### 2024-06-12 849e3ba
adding additional hints for contributions (#24)
