+++ doc/source/index.rst
@@ -33,5 +33,6 @@
    fri/doc/fri
    docs/kuka_documentation
    Paper <https://arxiv.org/pdf/2311.12709>
+   contributing
    Source Code <https://github.com/lbr-stack/lbr_fri_ros2_stack/tree/humble>
    Changelog <https://github.com/lbr-stack/lbr_fri_ros2_stack/releases>
