### 2024-06-24 #6138 @CameronDevine
For the record, the two outstanding issues for me are:

* lbr-stack/pyFRI#15
* lbr-stack/pyFRI#14
