+++ doc/source/contributing.rst
@@ -5,5 +5,5 @@ If you wish to contribute to any of the projects, please refer to the guidelines
 - `lbr_fri_ros2_stack <https://github.com/lbr-stack/lbr_fri_ros2_stack/blob/humble/.github/CONTRIBUTING.md>`_:octicon:`link-external`
 - `lbr_fri_idl <https://github.com/lbr-stack/lbr_fri_idl/blob/fri-1/.github/CONTRIBUTING.md>`_:octicon:`link-external`
 - `fri <https://github.com/lbr-stack/fri#contributing>`_:octicon:`link-external`
-- `pyFRI <https://github.com/lbr-stack/pyFRI/blob/main/.github/CONTRIBUTING.md>`_:octicon:`link-external`
+- `pyfri <https://github.com/lbr-stack/pyfri/blob/main/.github/CONTRIBUTING.md>`_:octicon:`link-external`
 - `lbr_stack_doc <https://github.com/lbr-stack/lbr_stack_doc/blob/main/.github/CONTRIBUTING.md>`_:octicon:`link-external`
