+++ .readthedocs.yml
@@ -12,8 +12,12 @@ build:
     - graphviz
   jobs:
     post_install:
+      # clone lbr_fri_ros2_stack, lbr_fri_idl, and fri through git via vcs
       - vcs import doc/source --input https://raw.githubusercontent.com/lbr-stack/lbr_fri_ros2_stack/humble/lbr_fri_ros2_stack/repos-fri-1.15.yaml
+      # unzip FRI-Client-SDK_Cpp.zip
       - unzip doc/source/fri/FRI-Client-SDK_Cpp.zip -d doc/source/fri/FRI-Client-SDK_Cpp
+      # clone pyfri (non-recursive, i.e. skip pybind11 and fri)
+      - git clone https://github.com/lbr-stack/pyfri.git doc/source/pyfri
 
 python:
    install:
