+++ doc/source/source_code.rst
@@ -1,0 +1,8 @@
+Source Code
+===========
+Source code for the corresponding packages:
+
+- ``fri``: `Source code <https://github.com/lbr-stack/fri>`_:octicon:`link-external`
+- ``lbr_fri_ros2_stack``: `Source code <https://github.com/lbr-stack/lbr_fri_ros2_stack>`_:octicon:`link-external`
+- ``lbr_fri_idl``: `Source code <https://github.com/lbr-stack/lbr_fri_idl>`_:octicon:`link-external`
+- ``pyfri``: `Source code <https://github.com/lbr-stack/pyfri>`_:octicon:`link-external`
