+++ .github/workflows/open_journals_pdf_generator.yml
@@ -18,7 +18,7 @@ jobs:
           journal: joss
           paper-path: paper/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v4
         with:
           name: paper
           path: paper/paper.pdf
