### 2024-09-23 #6138 @CameronDevine
It took longer than accepted, but I have now completed my review. Although I would have liked to test the code using the actual hardware, unsurprisingly, I don't own a Kuka IIWA. Therefore, I only have confirmed the functionality of the code using the simulation provided. Once, I was using the correct version of Ubuntu, the install went vary smoothly and the installation instructions and steps to run the examples were excellent. Therefore, I recommend that this paper be accepted.
