### 2024-10-01 #6138 @adi3
@mhubii While you're at it, could you also add the DOIs for the missing entries above? You can verify it anytime by running `@editorialbot check references`
