+++ paper/paper.md
@@ -99,7 +99,7 @@ Table: Overview of existing frameworks for interfacing the KUKA LBRs. A square i
 # Acknowledgement
 We want to acknowledge the work in [@iiwa_stack], as their MoveIt configurations were utilized in a first iteration of this project.
 
-This work was supported by core funding from the Wellcome/EPSRC [WT203148/Z/16/Z; NS/A000049/1]. This project has received funding from the European Union's Horizon 2020 research and innovation programme under grant agreement No 101016985 (FAROS project).
+This work was supported by core funding from the Wellcome/EPSRC [WT203148/Z/16/Z; NS/A000049/1], the European Union’s Horizon 2020 research and innovation programme under grant agreement No 101016985 (FAROS project), and EPSRC under the UK Government Guarantee Extension (EP/Y024281/1, VITRRO).
 
 # References
 <!-- compiled paper.bib through pandoc -->
