### 2024-10-07 #6138 @mhubii
awesome!

The references with no DOI do not have a DOI.

The releases are archived on Zenodo under DOI `10.5281/zenodo.13897377`.

Version: 2.0.0
