### 2024-10-21 #6138 @adi3
@crvernon can you please confirm if it's alright to have skipped DOIs in the references? Thanks! Apart from that this article looks all set to go out
