### 2024-11-14 #6138 @crvernon
:wave: @mhubii - I just need you to address the following before I move to accept this for publication:

In the paper:

- [ ] LINE 60:  "...the work in (Hennersperger et al., 2017)..." should appear as "...the work in Hennersperger et al. (2017)..."
- [ ] LINE 83:  The "g" in "gazebo" should be capitalized.  You can do this by wrapping any characters you want to preserve formatting of in your bib file using curly brackets.
- [ ] LINE 96:  "kuka" should be all caps.

Let me know when you have made these changes.  Thank you.
