### 2024-11-15 ba74ad3
title case citation style (https://github.com/openjournals/joss-reviews/issues/6138#issuecomment-2476338630)
