+++ paper/paper.bib
@@ -1,109 +1,120 @@
 @inproceedings{libiiwa,
-  author={Serrano-Muñoz, Antonio and Elguea-Aguinaco, Íñigo and Chrysostomou, Dimitris and BØgh, Simon and Arana-Arexolaleiba, Nestor},
-  booktitle={2023 IEEE/SICE International Symposium on System Integration (SII)}, 
-  title={A Scalable and Unified Multi-Control Framework for KUKA LBR iiwa Collaborative Robots}, 
-  year={2023},
-  volume={},
-  number={},
-  pages={1-5},
-  doi={10.1109/SII55687.2023.10039308}
+  author    = {Serrano-Muñoz, Antonio and Elguea-Aguinaco, Íñigo and Chrysostomou, Dimitris and BØgh, Simon and Arana-Arexolaleiba, Nestor},
+  booktitle = {2023 IEEE/SICE International Symposium on System Integration (SII)},
+  title     = {{A Scalable and Unified Multi-Control Framework for KUKA LBR IIWA Collaborative Robots}},
+  year      = {2023},
+  volume    = {},
+  number    = {},
+  pages     = {1-5},
+  doi       = {10.1109/SII55687.2023.10039308},
+  langid    = {english}
 }
 
 @software{iiwa_ros,
   author = {Chatzilygeroudis, Konstantinos and Mayr, Matthias and Fichera, Bernardo and Billard, Aude},
-  title = {iiwa_ros: A ROS Stack for KUKA's IIWA robots using the Fast Research Interface},
-  url = {http://github.com/epfl-lasa/iiwa_ros},
-  year = {2019},
+  title  = {{iiwa\_ros: A ROS Stack for KUKA’s IIWA Robots Using the Fast Research Interface}},
+  url    = {http://github.com/epfl-lasa/iiwa_ros},
+  year   = {2019},
+  langid = {english}
 }
 
-@ARTICLE{kuka_sunrise_toolbox,  
-  Author={M. {Safeea} and P. {Neto}},  
-  Journal={IEEE Robotics Automation Magazine},  
-  Title={KUKA Sunrise Toolbox: Interfacing Collaborative Robots With MATLAB},  
-  Year={2019},  
-  Volume={26},  
-  Number={1},  
-  Pages={91-96},  
-  doi={10.1109/MRA.2018.2877776},  
-  ISSN={1070-9932},  
-  Month={March},  
+@article{kuka_sunrise_toolbox,
+  author  = {M. {Safeea} and P. {Neto}},
+  journal = {IEEE Robotics Automation Magazine},
+  title   = {{KUKA Sunrise Toolbox: Interfacing Collaborative Robots With MATLAB}},
+  year    = {2019},
+  volume  = {26},
+  number  = {1},
+  pages   = {91-96},
+  doi     = {10.1109/MRA.2018.2877776},
+  issn    = {1070-9932},
+  month   = {March},
+  langid  = {english}
 }
 
 @misc{iiwa_ros2,
-  author = {Bednarczyk, Maciej and Guzmán, Juan Helios García},
-  title = {ROS 2 Stack for KUKA iiwa collaborative robots},
-  year = {2023},
+  author    = {Bednarczyk, Maciej and Guzmán, Juan Helios García},
+  title     = {{ROS 2 Stack for KUKA IIWA Collaborative Robots}},
+  year      = {2023},
   publisher = {GitHub},
-  journal = {GitHub repository},
-  url = {https://github.com/ICube-Robotics/iiwa_ros2}
+  journal   = {GitHub repository},
+  url       = {https://github.com/ICube-Robotics/iiwa_ros2},
+  langid    = {english}
 }
 
 @misc{ros2_control,
[…]
