### 2024-11-17 #6138 @CameronDevine
Congratulations @mhubii! I'm glad to see that this paper is finally accepted. I do have some feedback for the JOSS editors relating to this review. Where is the best place to share this feedback?
